Make notmuch search not omit excluded messages.
Call notmuch search with --exclude=false so that we get back information
about excluded messages. This allows us to selectively include messages
that would normally be excluded, without passing a list of tags for
notmuch search to not treat as exclude tags (as we did before).

We will also be able to display metadata about excluded messages.